first try at crossbeam::epoch
Store the Canary in an "Atomic".  What does this mean exactly?

We created a new Canary.  The Atomic stores a pointer to the Canary.

The Guard isn't doing much yet-- it will allow us to track the proper
lifetime of objects, but for now it's just boilerplate that allows us to
access the thing in the Atomic.

When we try to access the Canary, we do it via something called Shared.
A Shared<T> is presumably just a ref to the inner T, but it allows rust
to do some magic lifetime tracking: we just want to know that our access
to the T doesn't outlast the Guard.

std::sync::atomic::Ordering is a big topic, and maybe I'll cover that
later.  Consider it a placeholder for something I'll need to think about
once I know how that value will be accessed.

Note that the Atomic doesn't behave like a rust Box or other container,
in that it doesn't seem to manage the lifetime of the Canary.  This is
backed up by the fact that I no longer see the Canary's "dropped"
message.  Hopefully we'll learn how to deal with this soon.