play around with guard.defer
One of the critical parts about epoch-based memory management is the
concept of "deferred" work.  The main thing we want is probably deleting
released objects, but crossbeam will let you call arbitrary functions.

crossbeam::epoch puts the interesting calls into the Guard object, so
let's play around with Guard::defer().  It looks like the default
crossbeam behavior is to accumulate a lot of work on the local thread
before pushing that work to the global Collector, which will somehow
trigger the actual running of the deferred work.  There's a
Guard::flush() call to do that pushing immediately, if you're impatient.

It's also worth noting that in our small example, there in deferred work
that never runs-- our program simply ends.  I doubt that's a real
problem (who cares if we failed to free a bunch of data right before the
process exits), but maybe we'll figure that out later.