build a bigger birdcage
To really begin to exercise deferred memory reclamation, I need a bigger
data structure.  So let's make our BirdCage hold an arbitrary number of
Atomic<Canary> elements.

access() was about to get confusing because I used an integer for the
"context" string, so instead I'm switching that to a string so I can
have an integer argument specifying which element to access.

I had to fiddle with the cleanup() signature a bit, because the compiler
was throwing a bunch of lifetime errors.  Because we can't know when
cleanup() will actually run, any arguments to it must be by-value or
have 'static lifetime.  I'll copy the context string to make life
easier.