add a replace() method
If we're going to do concurrent operations requiring memory management,
we're going to need a way to actually remove things from our data
structure (BirdCage).

Add a replace() method, which goes into our data structure and removes
one Canary and replaces it with another.  We use Atomic::swap() to do
this.

Note we have removed the Canary pointer from the main data structure,
but we haven't actually done anything to free it.  Each time replace()
runs, it leaks one Canary.