add deferred destruction
When our data structure releases an item, we want it to be automatically
destroyed once the global epoch has advanced.  This could be used by
calling defer() and handing over the value to be destroyed, but that's
hard to do because you'd have to extract the pointed from the Shared
(which can't outlive the Guard).

Conveniently, there is a defer_destroy() function that does exactly what
we want.

As in previous examples, you need to run through a lot of pin()s and
accumulated deferred work before any of it actually happens; play with
the numbers (and try the flush() call) to see different effects.

There's a really strange bit of code here at the end, where I wanted to
see all of the deferred garbage collection happen before we exit.  I
couldn't find an easy way to do that though-- the best I could do was
create two more pins to force the epoch counter forward.

I don't think I understand the global epoch counter design well enough
to really explain why two calls is necessary, but it seems to me that
this sort of internal knowledge shouldn't really be required to do
something this basic.  Maybe there's a better way to do this, but I
haven't found it yet.