add random thread behavior
Now that we have what looks like a basic concurrent data structure, we
should actually add some threads to do random unpredictable things to
it.  We'd like to have callers to access() and replace() bumping into
each other constantly.

First: spawn a bunch of threads.  We need to be a little careful with
passing things to threads, since Rust can't reason about whether they
might outlive the main thread.  So we'll put the BirdCage in an Arc,
which will allow us to make thread-safe references.

We'll also number the threads so we can print out interesting messages
from each.

Finally, collect all the thread handles and join() them so main()
doesn't exit prematurely.

We'd like the threads to bang around randomly, so we'll also randomize
which locations each thread accesses.  Each thread will loop for a
while, doing repetitions of

    {
        access(random_index);
        replace(another_random_index);
    }

It's useful to crank up the number of threads, or iterations, or both.
Increasing the birdcage size will actually reduce contention, so try
making that smaller.  This won't prove that our data structure is
correct, but it's nice to see that it will run as long as you want
without crashing.

It's also worth checking the speed of debug builds vs release.  You
probably want to remove all the println! statements because the
different threads are spending most of their time fighting over the
stdout mutex and not doing real work.  The release build is about 10x
faster on my machine.